+++ src/main/java/com/nred/nuclearcraft/block/turbine/TurbineRedstonePortBlock.java
@@ -1,32 +1,22 @@
 package com.nred.nuclearcraft.block.turbine;
 
-import com.nred.nuclearcraft.block.GenericTooltipDeviceBlock;
+import com.nred.nuclearcraft.block.GenericActiveTooltipDeviceBlock;
 import com.nred.nuclearcraft.block.IActivatable;
-import it.zerono.mods.zerocore.lib.block.multiblock.IMultiblockPartType;
-import it.zerono.mods.zerocore.lib.multiblock.IMultiblockController;
+import com.nred.nuclearcraft.multiblock.turbine.ITurbinePartType;
+import com.nred.nuclearcraft.multiblock.turbine.Turbine;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.StateDefinition;
 import org.jetbrains.annotations.Nullable;
 
-import static com.nred.nuclearcraft.registration.BlockRegistration.ACTIVE;
-
-public class TurbineRedstonePortBlock<Controller extends IMultiblockController<Controller>, PartType extends IMultiblockPartType> extends GenericTooltipDeviceBlock<Controller, PartType> implements IActivatable {
-    public TurbineRedstonePortBlock(MultiblockPartProperties<PartType> properties) {
+public class TurbineRedstonePortBlock extends GenericActiveTooltipDeviceBlock<Turbine, ITurbinePartType> implements IActivatable {
+    public TurbineRedstonePortBlock(MultiblockPartProperties<ITurbinePartType> properties) {
         super(properties);
-        this.registerDefaultState(this.defaultBlockState().setValue(ACTIVE, false));
     }
 
     @Override
     public boolean canConnectRedstone(BlockState state, BlockGetter level, BlockPos pos, @Nullable Direction direction) {
         return direction != null;
-    }
-
-    @Override
-    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
-        builder.add(ACTIVE);
     }
 }